access: relocate decision-reason catalogue to access subdirectory
The flat-prefix directory rule maps each wyl_<domain>_* prefix
into its own subdirectory under wyrelog/. The decision-reason
catalogue (wyl_deny_reason_*) lives in the access policy domain,
so move the two files implementing it into wyrelog/access/
alongside future access-engine sources:

  wyrelog/wyl-access-decision-private.h -> wyrelog/access/decision-private.h
  wyrelog/wyl-access-decision.c         -> wyrelog/access/decision.c

Function and type names are unchanged; only file paths move. The
single internal include in decision.c switches from
"wyl-access-decision-private.h" to the sibling-relative
"decision-private.h", and tests/test-access-decision.c follows
the public include path "wyrelog/access/decision-private.h".

Verified locally on both build lanes:
  - default: 26/26 PASS, no behavioural change.
  - audit: 28/28 PASS.
